backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "localrag-backend"
version = "0.1.0"
description = "LocalRAG — local-first document intelligence backend"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sse-starlette>=2.1.0",
    "python-multipart>=0.0.9",

    # Config
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",

    # Storage
    "qdrant-client>=1.9.0",

    # Embeddings
    "sentence-transformers>=3.0.0",
    "FlagEmbedding>=1.2.0",
    "numpy>=1.26.0",

    # LLM Providers
    "httpx>=0.27.0",
    "openai>=1.35.0",
    "anthropic>=0.30.0",
    "google-generativeai>=0.7.0",

    # Document parsing
    "pymupdf>=1.24.0",
    "pdfplumber>=0.11.0",
    "python-docx>=1.1.0",
    "aiofiles>=23.2.0",

    # Utilities
    "loguru>=0.7.2",
    "tiktoken>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["."]
